% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corum_180627}
\alias{corum_180627}
\title{corum_180627}
\format{tibble}
\usage{
corum_180627
}
\description{
Tibble of the full CORUM protein complex database.
}
\details{
Source:

Ruepp A, Waegele B, Lechner M, Brauner B, Dunger-Kaltenbach I, Fobo G, Frishman G, Montrone C, Mewes HW. CORUM: the comprehensive resource of mammalian protein complexes—2009. Nucleic Acids Res. 2010 Jan;38(Database issue):D497-501. 19884131

Ruepp A, Brauner B, Dunger-Kaltenbach I, Frishman G, Montrone C, Stransky M, Waegele B, Schmidt T, Doudieu ON, Stümpflen V, Mewes HW. CORUM: the Comprehensive Resource of Mammalian Protein Complexes. Nucleic Acids Res. 2008 Jan;36(Database issue):D449-54 17965090
}
\keyword{datasets}
