% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corum_alt_ids_180627}
\alias{corum_alt_ids_180627}
\title{corum_alt_ids_180627}
\format{tibble}
\usage{
corum_alt_ids_180627
}
\description{
Alternate identifiers for complex subunits.
}
\keyword{datasets}
