% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corum_go_terms_180627}
\alias{corum_go_terms_180627}
\alias{corum_go_terms_180627}
\title{corum_go_terms_180627}
\format{tibble}
\usage{
corum_go_terms_180627

corum_go_terms_180627
}
\description{
GO Terms corresponding to complex subunits

Tidied CORUM data.
}
\keyword{datasets}
